Relocating cond
