+++ doc/releases/changelog-dev.md
@@ -46,6 +46,10 @@
 * Simplified the logic for re-arranging states before returning.
   [(#4817)](https://github.com/PennyLaneAI/pennylane/pull/4817)
 
+* Moved `qml.cond` and the `Conditional` operation from the `transforms` folder to the `ops/op_math` folder. `qml.transforms.Conditional`
+  will now be available as `qml.ops.Conditional`.
+  [(#)]()
+
 <h3>Breaking changes 💔</h3>
 
 * The `prep` keyword argument has been removed from `QuantumScript` and `QuantumTape`.
