+++ pennylane/ops/op_math/condition.py
@@ -15,9 +15,7 @@
 Contains the condition transform.
 """
 from functools import wraps
-from typing import Type
 
-from pennylane.measurements import MeasurementValue
 from pennylane.operation import AnyWires, Operation
 from pennylane.tape import make_qscript
 
@@ -48,12 +46,7 @@ class Conditional(Operation):
 
     num_wires = AnyWires
 
-    def __init__(
-        self,
-        expr: MeasurementValue[bool],
-        then_op: Type[Operation],
-        id=None,
-    ):
+    def __init__(self, expr, then_op, id=None):
         self.meas_val = expr
         self.then_op = then_op
         super().__init__(wires=then_op.wires, id=id)
